requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.0+cpu
torchvision==0.16.0+cpu
Pillow==10.0.1
python-multipart==0.0.6
numpy==1.24.4
scikit-learn==1.3.2
